Move common test classes to reactive-grpc-common
