+++ rx-java/rxgrpc-test/src/test/java/com/salesforce/rxgrpc/CancellationPropagationIntegrationTest.java
@@ -8,6 +8,7 @@
 package com.salesforce.rxgrpc;
 
 import com.google.protobuf.Empty;
+import com.salesforce.reactivegrpccommon.Sequence;
 import com.salesforce.servicelibs.NumberProto;
 import com.salesforce.servicelibs.RxNumbersGrpc;
 import io.grpc.*;
